refactor: Extracted module default_body
